client/requirements.txt
etcd-py==0.0.2
clint==0.3.2
nose==1.3.0
nose-progressive
